// src/app/modules/talent/interfaces/persona-info.ts
export interface PersonaInfo {
    ID:               string;
    first_name:       string;
    last_name:        string;
    gender:           string;
    birthdate:        Date;
    country:          string;
    state:            string;
    city:             string;
    move:             string;
    phone:            string;
    handicap:         string;
    travel:           string;
    civil_status:     string;
    linkedin:         string;
    github:           string;
    description:      string;
    has_car:          string;
    has_job:          string;
    license:          string;
    education:        string;
    profile_pic:      string;
    cv:               string;
    cv_visibility:    string;
    whatsapp:         string;
    contact_email:    string;
    nationality:      string;
    category:         string;
    complete_profile: string;
    updated_at:       Date;
}